Fixing import paths
