+++ crates/ibc-clients/ics07-tendermint/Cargo.toml
@@ -19,6 +19,12 @@ all-features = true
 # ibc dependencies
 ibc = { version = "0.47.0", path = "./../../ibc", default-features = false }
 ibc-client-tendermint-types = { version = "0.47.0", path = "./types", default-features = false }
+ibc-core-commitment-types = { version = "0.47.0", path = "./../../ibc-core/ics23-commitment/types" }
+ibc-core-client-context = { version = "0.47.0", path = "./../../ibc-core/ics02-client/context" }
+ibc-core-client-types = { version = "0.47.0", path = "./../../ibc-core/ics02-client/types" }
+ibc-core-context-types = { version = "0.47.0", path = "./../../ibc-core/context/types" }
+ibc-core-host-types = { version = "0.47.0", path = "./../../ibc-core/ics24-host/types" }
+ibc-primitives = { version = "0.47.0", path = "./../../ibc-primitives" }
 ibc-proto = { workspace = true, features = ["server"] }
 
 # cosmos dependencies
