+++ crates/ibc-clients/ics07-tendermint/types/src/consensus_state.rs
@@ -3,6 +3,7 @@
 use ibc_proto::google::protobuf::Any;
 use ibc_proto::ibc::lightclients::tendermint::v1::ConsensusState as RawConsensusState;
 use ibc_proto::Protobuf;
+
 use tendermint::hash::Algorithm;
 use tendermint::time::Time;
 use tendermint::Hash;
@@ -10,11 +11,11 @@ use tendermint_proto::google::protobuf as tpb;
 
 use crate::error::Error;
 use crate::header::Header;
-use ibc::core::ics02_client::consensus_state::ConsensusState as ConsensusStateTrait;
-use ibc::core::ics02_client::error::ClientError;
-use ibc::core::ics23_commitment::commitment::CommitmentRoot;
-use ibc::core::timestamp::Timestamp;
-use ibc::prelude::*;
+
+use ibc_core_client_context::consensus_state::ConsensusState as ConsensusStateTrait;
+use ibc_core_client_types::error::ClientError;
+use ibc_core_commitment_types::commitment::CommitmentRoot;
+use ibc_primitives::{prelude::*, Timestamp};
 
 pub const TENDERMINT_CONSENSUS_STATE_TYPE_URL: &str =
     "/ibc.lightclients.tendermint.v1.ConsensusState";
@@ -36,13 +37,19 @@ impl ConsensusState {
             next_validators_hash,
         }
     }
-    
-    pub fn timestamp(&self) -> Time {
-        self.timestamp
+}
+
+impl ConsensusStateTrait for ConsensusState {
+    fn root(&self) -> &CommitmentRoot {
+        &self.root
     }
 
-    pub fn root(&self) -> CommitmentRoot {
-        self.root.clone()
+    fn timestamp(&self) -> Timestamp {
+        self.timestamp.into()
+    }
+
+    fn encode_vec(self) -> Vec<u8> {
+        <Self as Protobuf<Any>>::encode_vec(self)
     }
 }
