+++ crates/ibc-primitives/src/utils/mod.rs
@@ -1,6 +1,6 @@
 //! Various utilities used internally
 mod macros;
-mod pretty;
+pub mod pretty;
 
 pub use macros::*;
 pub use pretty::*;
